Ex16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
 
 int main(void)
 {
-	long long int Fibo1 = 1;
-	long long int n = 0;
-	long long int Fibo2 = 1;
+	unsigned long int Fibo1 = 1;
+	unsigned long int n = 0;
+	unsigned long int Fibo2 = 1;
 	int m = 0;
 
 	while (m < 98)
